Separated different site parts into different files, added subtle CSS and JS animations, added functions that allow user to like playlists, fixed a bug, which denied the user the ability to add songs to playlists directly from search bar.
