+++ bottomTab.html
@@ -1,0 +1,10 @@
+<section id="bottomTab">
+    <button id="previous">&#9666;</button>
+    <button id="playBtn">&#9658;</button>
+    <button id="pauseBtn">&#10074;&#10074;</button>
+    <button id="next">&#9656;</button>
+    <h2 id="timer">00:00</h2>
+    <input type="range" min="0" max="100" step="1" id="seekSlider" value="0">
+    <h2>  &#128266;  </h2>
+    <input type="range" min="0" max="1" step="0.01" id="vol" value="0.2">
+</section>
