+++ leftTab.php
@@ -1,0 +1,58 @@
+<section id="leftTab">
+            <div id="fInputTitle">Select File</div>
+            <div id="fInput">
+                <form action="" method="post" id="f2" enctype="multipart/form-data">
+                    <?php 
+                        $x->formDisplay(); 
+                    ?>
+                </form>
+            </div>
+            <?php
+                $x->getSongData();
+            ?>
+            <div id="createPlaylistTitle">Create Playlist</div>
+            <div id="createPlaylist">
+                <form action="" method="post" id="f1" enctype="multipart/form-data">
+                    <label for="playlistName">Playlist name: </label>
+                    <input type="text" name="playlistName">
+                    <br>
+                    <br>
+                    <!--<label for="cover-art">Cover art: </label>
+                    <input type="file" name="cover-art">-
+                    <br>
+                    <br>-->
+                    <label>Playlist type: </label>
+                    <br>
+                    <div class="type">
+                        <label for="Public">Public</label>
+                        <div class="questionMark">(?)
+                            <span class="helpText">A public playlist will be visible to everyone</span>
+                        </div>
+                        <input type="radio" name="playlistType" value="Public">
+                    </div>
+                    <div class="type">
+                        <label for="Private">Private</label>
+                        <div class="questionMark">(?)
+                            <span class="helpText">A private playlist will only be visible to you and can only by accessed by you</span>
+                        </div>
+                        <input type="radio" name="playlistType" value="Private">
+                    </div>
+                    <div class="type">
+                        <label for="Local">Local</label>
+                        <div class="questionMark">(?)
+                            <span class="helpText">
+                                A local playlist is where you can store your <br>
+                                 own imported audio files from your device. <br>
+                                It will not be accessible to anyone and cannot be shared.
+                            </span>
+                        </div>
+                        <input type="radio" name="playlistType" value="Local">
+                    </div>
+                    <br>
+                    <button name="Create">Create</button>
+                </form>
+                <?php
+                    $x->getPlaylistData();
+                ?> 
+            </div>
+        </section>
